Address review: localized error copy and Long price values
- Use my_ads_load_failed for OlxApiException (its message is dev-facing)
- Switch OlxAdvertPrice.value from Float to Long to avoid precision loss

+++ composeApp/src/commonMain/kotlin/com/sirelon/aicalories/features/seller/auth/data/response/OlxAdvertResponse.kt
@@ -64,7 +64,7 @@ internal class OlxAdvertImageResponse(
 @Serializable
 internal class OlxAdvertPriceResponse(
     @SerialName("value")
-    val value: Float?,
+    val value: Long?,
 
     @SerialName("currency")
     val currency: String?,
